+++ src/ambition/version_details.cc
@@ -50,6 +50,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Hide icons for units that are dying."),
           _("Lined up the villager and peasant numbers on the Villages Report."),
           _("Increased the game speed of the unlimited speed by not drawing every frame, but only 60 FPS."),
+          _("Use thousands separators for food purchase quantities and aid and tribute amounts."),
         },
       },
     },